Update to 0.26.7
Submodule URLs and paths with a leading "-" are now ignored. This is due to the
recently discovered CVE-2018-17456, which can lead to arbitrary code execution
in upstream git. While libgit2 itself is not vulnerable, it can be used to
inject options in an implementation which performs a recursive clone by
executing an external command.

Full changelog available [here](https://github.com/libgit2/libgit2/blob/v0.26.7/CHANGELOG.md)

Signed-off-by: Peter O'Connor <peter@solus-project.com>